XNA - Phong model fully implemented with multiple light support (needs to be refactored for better structure)

+++ RayTracerXNA/RayTracerXNA/Primitive.cs
@@ -32,8 +32,8 @@ namespace RayTracerXNA
             set { specularStrength = value; }
         }
 
-        protected float exponent;
-        public float Exponent
+        protected double exponent;
+        public double Exponent
         {
             get { return exponent; }
             set { exponent = value; }
